refactor - move components to other files
